requirements.txt
python-telegram-bot==21.0.1
sentence-transformers
faiss-cpu
numpy
python-dotenv
protobuf